+++ grace-util/src/main/groovy/grails/util/Pair.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2022 the original author or authors.
+ * Copyright 2013-2025 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,18 +15,15 @@
  */
 package grails.util;
 
-public class Triple<A, B, C> {
+public class Pair<A, B> {
 
     final A aValue;
 
     final B bValue;
 
-    final C cValue;
-
-    public Triple(A aValue, B bValue, C cValue) {
+    public Pair(A aValue, B bValue) {
         this.aValue = aValue;
         this.bValue = bValue;
-        this.cValue = cValue;
     }
 
     public A getaValue() {
@@ -37,16 +34,11 @@ public class Triple<A, B, C> {
         return this.bValue;
     }
 
-    public C getcValue() {
-        return this.cValue;
-    }
-
     @Override
     public int hashCode() {
         int result = 1;
         result = 31 * result + ((this.aValue == null) ? 0 : this.aValue.hashCode());
         result = 31 * result + ((this.bValue == null) ? 0 : this.bValue.hashCode());
-        result = 31 * result + ((this.cValue == null) ? 0 : this.cValue.hashCode());
         return result;
     }
 
@@ -61,7 +53,7 @@ public class Triple<A, B, C> {
         if (getClass() != obj.getClass()) {
             return false;
         }
-        Triple other = (Triple) obj;
+        Pair other = (Pair) obj;
         if (this.aValue == null) {
             if (other.aValue != null) {
                 return false;
@@ -71,27 +63,16 @@ public class Triple<A, B, C> {
             return false;
         }
         if (this.bValue == null) {
-            if (other.bValue != null) {
-                return false;
-            }
+            return other.bValue == null;
         }
-        else if (!this.bValue.equals(other.bValue)) {
-            return false;
+        else {
+            return this.bValue.equals(other.bValue);
         }
-        if (this.cValue == null) {
-            if (other.cValue != null) {
-                return false;
-            }
-        }
-        else if (!this.cValue.equals(other.cValue)) {
-            return false;
-        }
-        return true;
     }
 
     @Override
     public String toString() {
-        return "Triple [aValue=" + this.aValue + ", bValue=" + this.bValue + ", cValue=" + this.cValue + "]";
+        return "TupleKey [aValue=" + this.aValue + ", bValue=" + this.bValue + "]";
     }
 
 }
