+++ grace-util/src/main/groovy/grails/util/TypeConvertingMap.groovy
