+++ grace-util/src/test/groovy/grails/util/TypeConvertingMapTests.groovy
